Safari-611.1.10/LayoutTests/crypto/subtle/digest-malformed-parameters.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test digest with malformed parameters");

var message = asciiToUint8Array("Hello, World!");

// Not enough arguments.
shouldReject('crypto.subtle.digest()');
shouldReject('crypto.subtle.digest(1)');

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
